4.21

+++ part4/controllers/blogs.js
@@ -38,8 +38,20 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if ( user && blog && blog.user.toString() === user._id.toString() ) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(201).end()
+    }
+    else {
+        return response.status(401).json({ error: 'no permission' })
+    }
 })
 
 blogRouter.put('/:id', async (request, response) => {
